Fixing class format
